Added Time Played log + Install/Uninstall bug fixes
Added the game time tracking feature. Once the game is launched it will track in seconds how long it was running for before appending it in the config.ini.

Also tried fixing some more bugs in the install and uninstall functions. Can still be problematic so will need to take another look.