Fixed modlist when compiled
Fixed an issue where the modlist was not readable when 2KAN was packaged into an exe using auto-py-to-exe.

The fix (albeit a pretty terrible one) is to simply pass the modlist.json full path to every function that needs it from the config file. Yes I know, its a terrible way of doing it, at this point I'm over this codebase and framework and just want something that works.