Fixed Spacedock API + New Logo + Other fixes
Fixed the spacedock API so it now grabs all pages returned via GET request. A new logo courtesy of Midjourney (thank you AI overlords). Also fixed some formatting issues in the header and modlist as well as removing "Top" as an option in the available mods menu as it was not returning anything.